@@ -44,11 +44,11 @@
     <properties>
         <timestamp>${maven.build.timestamp}</timestamp>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
 
         <!-- Dependencies -->
-        <spring.boot.version>2.7.0</spring.boot.version>
-        <beanmapper.version>3.1.1</beanmapper.version>
+        <spring.boot.version>3.0.5</spring.boot.version>
+        <beanmapper.version>5.0.0</beanmapper.version>
     </properties>
 
     <dependencyManagement>
