@@ -3,20 +3,17 @@ package nl._42.beanie.spring;
 import nl._42.beanie.BeanBuilder;
 import nl._42.beanie.spring.domain.Person;
 import nl._42.beanie.spring.domain.PersonResult;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.function.Function;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 @SpringBootTest
-@RunWith(SpringRunner.class)
 public class ConfigurationTest {
 
     @Autowired
