@@ -3,8 +3,8 @@
  */
 package nl._42.beanie.generator.increment;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 /**
  * 
@@ -17,9 +17,9 @@ public class IncrementingIntegerGeneratorTest {
     @Test
     public void testSequence() {
         IntegerValueIncrementor generator = new IntegerValueIncrementor();
-        Assert.assertEquals(1, generator.generate(null));
-        Assert.assertEquals(2, generator.generate(null));
-        Assert.assertEquals(3, generator.generate(null));
+        Assertions.assertEquals(1, generator.generate(null));
+        Assertions.assertEquals(2, generator.generate(null));
+        Assertions.assertEquals(3, generator.generate(null));
     }
 
 }
