@@ -1,25 +1,26 @@
 package nl._42.beanie.support;
 
 import nl._42.beanie.util.Classes;
-
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class ClassesTest {
 	
 	@Test
 	public void testForName() {
-		Assert.assertEquals(getClass(), Classes.forName(getClass().getName()));
+		Assertions.assertEquals(getClass(), Classes.forName(getClass().getName()));
 	}
 
-	@Test(expected = RuntimeException.class)
+	@Test
 	public void testForNameNonExisting() {
-		Classes.forName("my.unknown.Class");
+		Assertions.assertThrows(RuntimeException.class, () ->
+			Classes.forName("my.unknown.Class")
+		);
 	}
 	
 	@Test
 	public void testHasNullaryConstructor() {
-		Assert.assertTrue(Classes.hasNullaryConstructor(getClass()));
+		Assertions.assertTrue(Classes.hasNullaryConstructor(getClass()));
 	}
 	
 }
