@@ -7,7 +7,7 @@ public class MockitoBeanGenerator implements ValueGenerator {
 	@Override
 	public Object generate(Class<?> valueType) {
 		Object mock = Mockito.mock(valueType);
-		Mockito.verifyZeroInteractions(mock);
+		Mockito.verifyNoInteractions(mock);
 		return mock;
 	}
 	
